Move "statistics" route into user folder
Before this change, only members and companies could access the "statistics" view. This commit moves the "statistics" view into the user folder, deleting thereby code duplication and making it acessible to accountants as well.
