+++ arbeitszeit_flask/templates/company/dashboard.html
@@ -133,7 +133,7 @@
             <div class="tile is-parent"></div>
         </div>
         <div class="tile is-ancestor">
-            <a class="tile is-parent" href="{{ url_for('main_company.statistics') }}">
+            <a class="tile is-parent" href="{{ url_for('main_user.statistics') }}">
                 <div class="tile is-child box has-background-warning-light">
                     <h1 class="title is-5"><span class="icon"><i class="fa-solid fa-chart-pie"></i></span> {{
                         gettext("Global statistics") }}
