+++ arbeitszeit_flask/templates/member/dashboard.html
@@ -99,7 +99,7 @@
       <div class="tile is-parent"></div>
     </div>
     <div class="tile is-ancestor">
-      <a class="tile is-parent" href="{{ url_for('main_member.statistics') }}">
+      <a class="tile is-parent" href="{{ url_for('main_user.statistics') }}">
         <div class="tile is-child box has-background-warning-light">
           <h1 class="title is-5"><span class="icon"><i class="fa-solid fa-chart-pie"></i></span> {{
             gettext("Global statistics") }}
