--- arbeitszeit_flask/templates/member/statistics.html
@@ -1,12 +1,0 @@
-{% extends "base.html" %}
-
-{% block navbar_start %}
-<div class="navbar-item">{{ gettext("Global statistics") }}</div>
-{% endblock %}
-
-{% block content %}
-
-{% from 'macros/statistics.html' import statistics %}
-{{ statistics(view_model) }}
-
-{% endblock %}
