+++ arbeitszeit_flask/user/routes.py
@@ -15,6 +15,7 @@ from arbeitszeit_flask.class_based_view import as_flask_view
 from arbeitszeit_flask.forms import RequestEmailAddressChangeForm
 from arbeitszeit_flask.types import Response
 from arbeitszeit_flask.views import QueryCompaniesView, QueryPlansView
+from arbeitszeit_flask.views.get_statistics_view import GetStatisticsView
 from arbeitszeit_flask.views.http_error_view import http_404, http_501
 from arbeitszeit_web.www.controllers.request_email_address_change_controller import (
     RequestEmailAddressChangeController,
@@ -105,4 +106,10 @@ class query_plans(QueryPlansView):
 @AuthenticatedUserRoute("/query_companies", methods=["GET"])
 @as_flask_view()
 class query_companies(QueryCompaniesView):
+    ...
+
+
+@AuthenticatedUserRoute("/statistics")
+@as_flask_view()
+class statistics(GetStatisticsView):
     ...
